refactored project into sub-projects
