+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = "i-eat-what-backend"
-include 'models'
+include 'src:models', 'src:api'
