+++ src/api/build.gradle
@@ -1,0 +1,6 @@
+apply plugin : 'java'
+
+dependencies {
+    compile project(':src:models')
+    compile ratpack.dependency("hikari")
+}
