--- src/models/java/Models.java
@@ -1,6 +1,0 @@
-public class Models {
-
-    static void main(String... args) {
-        System.out.println("Test Models");
-    }
-}
