+++ src/models/src/main/java/Article.java
@@ -1,40 +1,52 @@
-import com.fasterxml.jackson.databind.ObjectMapper;
-import ratpack.handling.Context;
-import ratpack.handling.Handler;
+public class Article {
 
-import java.util.HashMap;
-import java.util.Map;
+    private String id;
+    private String title;
+    private String blurb;
+    private ArticleSection[] sections;
 
-public class PublicAPI {
+    public Article(){}
 
-    public Map<String, Handler> postRequestHandlers() {
-        Map<String, Handler> handlers = new HashMap<>();
-
-        return handlers;
+    public Article(String id, String title, String blurb, ArticleSection... articleSections) {
+        setId(id);
+        setTitle(title);
+        setBlurb(blurb);
+        setSections(articleSections);
     }
 
-    public Map<String, Handler> getRequestHandlers() {
-        Map<String, Handler> handlers = new HashMap<>();
-        handlers.put("", (ctx) -> ctx.render("FOO BAT METAL"));
-        handlers.put("public", (ctx) -> ctx.render("Public API Dynamic Success!"));
-        handlers.put(
-            "demo",
-            (ctx) -> {
-                ObjectMapper om = new ObjectMapper();
-//                final JsonNodeFactory jsonNodeFactory = JsonNodeFactory.instance;
-//                ObjectNode articleJSON = jsonNodeFactory.objectNode();
-                Article article = dummyContent();
-                String articleAsString = om.writeValueAsString(article);
-                ctx.render(articleAsString);
-            });
-        return handlers;
+    public String getId() {
+        return id;
     }
 
-    private void eventHandler(Context ctx) {
-
+    public String getTitle() {
+        return title;
     }
 
-    public static Article dummyContent() {
+    public String getBlurb() {
+        return blurb;
+    }
+
+    public ArticleSection[] getSections() {
+        return sections;
+    }
+
+    private void setId(String id) {
+        this.id = id;
+    }
+
+    private void setBlurb(String blurb) {
+        this.blurb = blurb;
+    }
+
+    private void setTitle(String title) {
+        this.title = title;
+    }
+
+    private void setSections(ArticleSection[] sections) {
+        this.sections = sections;
+    }
+
+    public static Article demoData() {
         ArticleSection[] sections;
         sections = new ArticleSection[3];
         String paragraph = "Zombie ipsum reversus ab viral inferno, nam rick grimes malum cerebro. De carne lumbering animata corpora quaeritis. Summus brains sit\\u200B\\u200B, morbo vel maleficia? De apocalypsi gorger omero undead survivor dictum mauris. Hi mindless mortuis soulless creaturas, imo evil stalking monstra adventus resi dentevil vultus comedat cerebella viventium. Qui animated corpse, cricket bat max brucks terribilem incessu zomby. The voodoo sacerdos flesh eater, suscitat mortuos comedere carnem virus. Zonbi tattered for solum oculi eorum defunctis go lum cerebro. Nescio brains an Undead zombies. Sicut malus putrid voodoo horror. Nigh tofth eliv ingdead. Zombie ipsum reversus ab viral inferno, nam rick grimes malum cerebro. De carne lumbering animata corpora quaeritis. Summus brains sit\\u200B\\u200B, morbo vel maleficia? De apocalypsi gorger omero undead survivor dictum mauris. Hi mindless mortuis soulless creaturas, imo evil stalking monstra adventus resi dentevil vultus comedat cerebella viventium. Qui animated corpse, cricket bat max brucks terribilem incessu zomby. The voodoo sacerdos flesh eater, suscitat mortuos comedere carnem virus. Zonbi tattered for solum oculi eorum defunctis go lum cerebro. Nescio brains an Undead zombies. Sicut malus putrid voodoo horror. Nigh tofth eliv ingdead.";
@@ -44,4 +56,3 @@ public class PublicAPI {
         return new Article("1235ASD", "Zombie Ipsum", "Brains!!!!", sections);
     }
 }
-
