+++ src/models/src/main/java/ArticleSection.java
